refactor: export server code from digitraffic/server
exporting from index breaks client components as they can't import node specific modules
