+++ site/src/lib/digitraffic/index.ts
@@ -1,5 +1,3 @@
 export * from './types'
 export * from './hooks'
 export * from './utils'
-
-export * from './get_stations'
